+++ src/main/java/com/oop/napi/domain/model/Usuario.java
@@ -7,7 +7,7 @@ import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Size;
 
-@Entity
+@Entity(name = "Usuario")
 public class Usuario {
 
     @Id
