+++ src/main/resources/application.properties
@@ -1,7 +1,10 @@
 spring.datasource.url=jdbc:mysql://localhost/napidb?createDatabaseIfNotExist=true&serverTimezone=UTC
 spring.datasource.username=root
-spring.datasource.password=123456
+spring.datasource.password=Hotdog12!br0
 
 spring.jpa.show-sql=true
+
+spring.flyway.baseline-on-migrate=true
+spring.flyway.locations=classpath:db.migration
 
 jwt.secret=usersecret
