+++ src/main/resources/db.migration/V0001__cria-tabela-usuario.sql
@@ -1,0 +1,8 @@
+CREATE TABLE Usuario (
+    id bigint not null auto_increment,
+    nome varchar(60) not null,
+    email varchar(255) not null,
+    senha varchar(255) not null,
+
+    primary key (id)
+)
