+++ src/main/resources/db.migration/V0002__cria-tabela-personagem.sql
@@ -1,0 +1,12 @@
+CREATE TABLE Personagem (
+    id bigint not null auto_increment,
+    nome varchar(60) not null,
+    sexo varchar(255) not null,
+    cla varchar(255) not null,
+    aldeia varchar(255) not null,
+    registroNinja varchar(255),
+    patenteNinja varchar(255),
+    primary key (id),
+
+    CONSTRAINT verificar_sexo CHECK (sexo='M' OR sexo='F')
+)
