+++ src/main/resources/db.migration/V0003__cria-tabela-estilo.sql
@@ -1,0 +1,5 @@
+CREATE TABLE Estilo (
+    id bigint not null,
+    estilo varchar(255) not null,
+    FOREIGN KEY (id) REFERENCES Personagem(id)
+)
