+++ src/main/resources/db.migration/V0004__popula-tabela-personagem.sql
@@ -1,0 +1,6 @@
+INSERT INTO Personagem (nome, sexo, cla, aldeia, registroNinja, patenteNinja) VALUES ('Naruto', 'M', 'Uzumaki', 'Folha', '12607', 'Kage');
+INSERT INTO Personagem (nome, sexo, cla, aldeia, registroNinja, patenteNinja) VALUES ('Sasuke', 'M', 'Uchiha', 'Folha', '12606', 'Genin');
+INSERT INTO Personagem (nome, sexo, cla, aldeia, registroNinja, patenteNinja) VALUES ('Sakura', 'F', 'Haruno', 'Folha', '12601', 'Chunin');
+INSERT INTO Personagem (nome, sexo, cla, aldeia, registroNinja, patenteNinja) VALUES ('Gaara', 'M', 'Clã do Kazekage', 'Areia', '56-001', 'Kage');
+INSERT INTO Personagem (nome, sexo, cla, aldeia, registroNinja, patenteNinja) VALUES ('Temari', 'F', 'Clã do Kazekage', 'Areia', '53-004', 'Jonin');
+INSERT INTO Personagem (nome, sexo, cla, aldeia, registroNinja, patenteNinja) VALUES ('Kankuro', 'M', 'Clã do Kazekage', 'Areia', '54-002', 'Jonin');
