+++ src/main/resources/db.migration/V0005__popula-tabela-estilo.sql
@@ -1,0 +1,2 @@
+insert into Estilo(id, estilo) values (1, 'Rasengan');	insert into Estilo(id, estilo) values (1, 'Clones das Sombras');	insert into Estilo(id, estilo) values (1, 'Jutsu Sexy');
+insert into Estilo(id, estilo) values (2, 'Chidori');	insert into Estilo(id, estilo) values (2, 'Susanoo');	insert into Estilo(id, estilo) values (2, 'Mangekyo Sharingan');
